Create README.md
Chatbot creation using Dialogflow
Websites play a big role in the conversion of potential customers to sales. Businesses have realized 
that with the addition of chatbots to their webpages, visitors stay engaged for much longer, thereby 
greatly improving these conversion rates.
Chatbots have also been acknowledged as an excellent resource for collecting and sharing relevant 
information. Furthermore, the automation of simple business processes without sacrificing human 
resources makes this a very economical way of generating online value.
Dialogflow : Dialogflow is a Google-owned framework that enables users to develop humancomputer interaction technologies that can support Natural Language Processing (NLP). Basically, it 
lets you make Digital Programs that interact with end users through natural languages. Therefore, 
you could even say that Dialogflow enables the creation of Conversational User Experience 
Platforms (CUXP)
Text/Voice – These are the modes used to communicate the input or the output. The user interacts 
with the bot through text or through voice. Text would be anything that is typed into the chatbot 
window and voice would be any message spoken into the chatbot window.
Agent – An agent is merely another term used to refer to the chatbot. Sometimes people say ‘agent’ 
when referring to the processing module within the application that enables discussions with the 
chatbot. And sometimes, it is another way to refer to the bot since it functions ‘like a support agent’. 
The context will always be clear enough for you know what they mean.
Expressions – Expressions/Training Phrases are the dialogues/utterances that people say when they 
interact with a bot. They represent a user’s desire and are often in the form of a question. For 
example –
“Is the store open?”
“Do you serve vegetarian?